Move ProjectVerifier to bitbucket.server
