--- auth/bitbucket/project_config.go
@@ -1,11 +1,0 @@
-package bitbucket
-
-type ProjectConfig struct {
-	Name string `json:"project_name,omitempty"`
-}
-
-func (flag *ProjectConfig) UnmarshalFlag(value string) error {
-	flag.Name = value
-
-	return nil
-}
